backend/requirements.txt

Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
requests==2.32.3
pillow==11.3.0
google-generativeai==0.8.5
python-dotenv==1.0.0
google-generativeai==0.8.5
python-dotenv==1.0.0